Refactor Change Printing dialog per review feedback
- Lift the menu entry out of the section switch in CEditorConstructed.
  addChangePrintingEntryIfApplicable now runs once after the switch
  (mirroring addMakeFoils), so Vanguard, Schemes, and Planes — which
  can also have alternative printings — get the entry when applicable.
  The internal predicate (infinite catalog + selected card + multiple
  printings) still gates display, so non-applicable sections hide it.

- Rename "Change Art..." -> "Change printing..." across desktop and
  mobile. Class ChangeArtDialog -> ChangePrintingDialog, lblChangeArt*
  -> lblChangePrinting*. Mobile FDeckEditor consolidated onto the
  shared lblChangePrinting / lblSelectPrinting keys; the unused
  lblChangePreferredArt / lblSelectPreferredArt entries pruned from
  en-US and the eight non-en-US locale files.

- Extract forge.toolbox.special.CardImageGrid as a reusable
  IDisposable Swing widget. Owns the JList HORIZONTAL_WRAP layout
  with size-override workarounds, the cell renderer, the HiDPI
  BICUBIC icon, the per-instance icon cache, the FScrollPane
  wrapping, and ImageFetcher placeholder handling. Filter UX
  deliberately not baked in — callers compose their own search bar
  over setItems. ChangePrintingDialog keeps the deck-art-specific
  bits (ArtStyle enum, search bar, debounce, empty-state swap,
  performSwap).
