+++ forge-gui/res/languages/de-DE.properties
@@ -1166,8 +1166,6 @@ lblNameNewDeck=Gib einen Namen ein
 lblSaveChangesCurrentDeck=Änderungen sichern?
 lblAddFavorites=Zu den Favoriten hinzufügen
 lblRemoveFavorites=Von den Favoriten entfernen
-lblChangePreferredArt=Setze bevorzugte Form
-lblSelectPreferredArt=Wähle bevorzugte Form für
 lblReplaceCard=Ersetze Karte
 lblReplace=Ersetze
 lblSelectReplacementCard=Wähle Ersatzkarte für
