+++ forge-gui/res/languages/en-US.properties
@@ -1243,9 +1243,9 @@ lblNameNewDeck=Enter name for new deck
 lblSaveChangesCurrentDeck=Save changes to current deck?
 lblAddFavorites=Add to Favorites
 lblRemoveFavorites=Remove from Favorites
-lblChangeArt=Change Art...
-lblChangeArtDialogTitle=Change Art — {0}
-lblChangeArtSearchHint=Filter by set name or code...
+lblChangePrinting=Change printing...
+lblChangePrintingDialogTitle=Change printing — {0}
+lblChangePrintingSearchHint=Filter by set name or code...
 lblArtStyleAll=All Styles
 lblArtStyleStandard=Standard
 lblArtStyleBorderless=Borderless
@@ -1254,9 +1254,8 @@ lblArtStyleShowcase=Showcase
 lblArtStyleExtendedArt=Extended Art
 lblArtStyleRetroFrame=Retro Frame
 lblArtStylePromo=Promo
-lblChangeArtNoResults=No printings match the current filter.
-lblChangePreferredArt=Change Preferred Art
-lblSelectPreferredArt=Select preferred art for
+lblChangePrintingNoResults=No printings match the current filter.
+lblSelectPrinting=Select printing for
 lblReplaceCard=Replace Card Variant
 lblReplace=Replace
 lblSelectReplacementCard=Select Replacement Card Variant for
