+++ forge-gui/res/languages/es-ES.properties
@@ -1145,8 +1145,6 @@ lblNameNewDeck=Introduce un nombre para el nuevo mazo
 lblSaveChangesCurrentDeck=¿Guardar los cambios al mazo actual?
 lblAddFavorites=Añadir a favoritos
 lblRemoveFavorites=Eliminar de favoritos
-lblChangePreferredArt=Cambiar arte preferido
-lblSelectPreferredArt=Selecciona el arte preferido para
 lblReplaceCard=Reemplazar Carta
 lblReplace=Reemplazar
 lblSelectReplacementCard=Selecciona la carta de reemplazo
