+++ forge-gui/res/languages/fr-FR.properties
@@ -1143,8 +1143,6 @@ lblNameNewDeck=Entrez le nom du nouveau deck
 lblSaveChangesCurrentDeck=Enregistrer les modifications dans le deck actuel ?
 lblAddFavorites=Ajouter aux favoris
 lblRemoveFavorites=Supprimer des favoris
-lblChangePreferredArt=Changer l'illustration préférée
-lblSelectPreferredArt=Sélectionner l'art préféré pour
 lblReplaceCard=Remplacer la variante de la carte
 lblReplace=Remplacer
 lblSelectReplacementCard=Sélectionner une variante de carte de remplacement pour
