+++ forge-gui/res/languages/it-IT.properties
@@ -1140,8 +1140,6 @@ lblNameNewDeck=Inserisci un nome per il nuovo mazzo
 lblSaveChangesCurrentDeck=Salvare le modifiche al mazzo corrente?
 lblAddFavorites=Aggiungi ai preferiti
 lblRemoveFavorites=Rimuovi dai preferiti
-lblChangePreferredArt=Modifica l'arte preferita
-lblSelectPreferredArt=Seleziona l'arte preferita per
 lblReplaceCard=Replace Card
 lblReplace=Replace
 lblSelectReplacementCard=Select Replacement Card for
