+++ forge-gui/res/languages/ja-JP.properties
@@ -1142,8 +1142,6 @@ lblNameNewDeck=新しいデッキの名前を入力
 lblSaveChangesCurrentDeck=現在のデッキへの変更を保存しますか？
 lblAddFavorites=お気に入りに追加
 lblRemoveFavorites=お気に入りから削除
-lblChangePreferredArt=優先アートの変更
-lblSelectPreferredArt=優先アートを選択
 lblReplaceCard=Replace Card
 lblReplace=Replace
 lblSelectReplacementCard=Select Replacement Card for
