+++ forge-gui/res/languages/ko-KR.properties
@@ -1188,8 +1188,6 @@ lblNameNewDeck=새로운 덱의 이름을 입력하세요
 lblSaveChangesCurrentDeck=현재 덱에 대한 변경 사항을 저장하시겠습니까
 lblAddFavorites=즐겨찾기에 추가
 lblRemoveFavorites=즐겨찾기에서 삭제
-lblChangePreferredArt=우선 아트 변경
-lblSelectPreferredArt=우선 아트 선택
 lblReplaceCard=카드 교체
 lblReplace=교체
 lblSelectReplacementCard=교체할 카드 선택 (대상)
