+++ forge-gui/res/languages/pt-BR.properties
@@ -1167,8 +1167,6 @@ lblNameNewDeck=Digite o nome do novo deck
 lblSaveChangesCurrentDeck=Salvar mudanças no deck atual?
 lblAddFavorites=Adicionar aos Favoritos
 lblRemoveFavorites=Remover dos Favoritos
-lblChangePreferredArt=Alterar Arte Preferida
-lblSelectPreferredArt=Selecionar arte preferida para
 lblReplaceCard=Substituir Variante de Carta
 lblSelectReplacementCard=Selecionar Variante de Carta Substituta para
 lblTo=para
