+++ forge-gui/res/languages/zh-CN.properties
@@ -1146,8 +1146,6 @@ lblNameNewDeck=输入新套牌的名称
 lblSaveChangesCurrentDeck=保存当前更改到套牌？
 lblAddFavorites=添加到收藏夹
 lblRemoveFavorites=从收藏夹删除
-lblChangePreferredArt=改变首选卡图
-lblSelectPreferredArt=选择首选的卡图版本
 lblReplaceCard=替代卡
 lblReplace=Replace
 lblSelectReplacementCard=选择替代卡
